fix: stackoverflow
